% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_specified_time_period.R
\name{extract_data_specified_time_period}
\alias{extract_data_specified_time_period}
\title{指定した時間帯のデータを抜き出す}
\usage{
extract_data_specified_time_period(df, datetime_col_name, start, end,
  start_equal = TRUE, end_equal = TRUE)
}
\arguments{
\item{df}{データフレーム}

\item{datetime_col_name}{日時情報が格納された列名}

\item{start}{開始時間}

\item{end}{終了時間}

\item{start_equal}{開始時間を含めるかどうか}

\item{end_equal}{終了時間を含めるかどうか}
}
\value{
指定した時間帯のみを抜き出したデータ
}
\description{
指定した時間帯のデータを抜き出す
}
\examples{
extract_data_specified_time_period(df1, Datetime, "2017-10-12 05:00", "2017-10-12 05:30", TRUE, TRUE)
extract_data_specified_time_period(df2, JST, "2017-10-12 05:00", "2017-10-12 05:30", False, TRUE)
}
