% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunrise_sunset.R
\name{get_sunrise_sunset}
\alias{get_sunrise_sunset}
\title{日付および緯度経度を指定し、日の出と日の入りの時間を取得する}
\usage{
get_sunrise_sunset(year, month, day, lat, lng)
}
\arguments{
\item{year}{年}

\item{month}{月}

\item{day}{日}

\item{lat}{緯度}

\item{lon}{経度}
}
\value{
日の出の時間と日の入りの時間
}
\description{
日付および緯度経度を指定し、日の出と日の入りの時間を取得する
}
\examples{
get_sunrise_sunset(2017, 1, 13, 45.1 ,135.4)
get_sunrise_sunset(2010, 3, 31, 25.1 ,138.4)
}
