% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attr_day_night.R
\name{set_attr_day_night}
\alias{set_attr_day_night}
\title{与えられたデータフレームにDayNightという列名にDay/Night/Sunset/Sunriseのカテゴリを付与}
\usage{
set_attr_day_night(df, datetime_col_name, lat, lng, tz,
  datetime_format = "\%Y-\%m-\%d \%H:\%M:\%S", sr_befor = 15,
  sr_after = 15, ss_before = 15, ss_after = 15)
}
\arguments{
\item{df}{データフレーム}

\item{datetime_col_name}{日時情報が格納された列名}

\item{lat}{緯度}

\item{lng}{経度}

\item{tz}{タイムゾーン}

\item{datetime_format}{時間の列名のフォーマット}

\item{sr_befor}{日の出の何分前までをsunsetとするか}

\item{sr_after}{日の出の何分後までをsunsetとするか}

\item{ss_before}{日の出の何分前までをsunriseとするか}

\item{ss_after}{日の出の何分後までをsunriseとするか}
}
\value{
DayNightという列名にday/night/sunset/sunriseのカテゴリを付与したデータフレーム
}
\description{
与えられたデータフレームにDayNightという列名にDay/Night/Sunset/Sunriseのカテゴリを付与
}
\examples{
set_attr_day_night(df1, Datetime, 45.1 ,135.4)
set_attr_day_night(df1, Datetime, 45.1 ,135.4, "Asia/Tokyo", "\%Y/\%m/\%d \%H:\%M", 0, 30, 15, 15)
}
